New org names, table, map
Suggest new org names based on feedback. It was suggested that the Table 2 and Figure 3 should match better, and that the data source should either be an individual contribution, or a repository/collection, which is indicated at the highest level possible (e.g., "MGDS" rather than "SOI", "FathomNet" rather than "MBARI", etc). We could break these down further in Appendices like has been done for Pangaea, NOAA, AADC, etc. 

I left the merged SQUIDLE/IMOS and IMOS because I think some of those were not from SQUIDLE? RLS was all from SQUIDLE, correct?

I assigned the same colors to different branches of the same group (e.g., DFO (IOS), DFO (BIO)). I didn't try to mess with the map legend, but I'm guessing we may need to adjust that to remove duplication of legend items.

"Merlin Best" was commented as a "Research Group", but it should be under "DFO (IOS)", which can just be "DFO" for the map. I think maybe this change needs to be implemented in the CSV, but I wasn't sure.